pyproject.toml
[tool.poetry]
name = "ticket_selling"
version = "0.1.0"
description = "Small application that provides API for tickets booking."
authors = ["Robert Strzybny"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.3"
djangorestframework = "^3.12.2"
django-celery-beat = "^2.1.0"
celery = "4.4.4"
future = "^0.18.2"
django-celery-results = "^1.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
